Adjust gatsby example
