--- examples/gatsby-functions/.nowignore
@@ -1,2 +1,0 @@
-yarn.lock
-README.md
