--- examples/gatsby-functions/gatsby-config.js
@@ -1,17 +1,0 @@
-module.exports = {
-  siteMetadata: {
-    title: 'Gatsby + Node.js (TypeScript) API',
-  },
-  plugins: [
-    `gatsby-plugin-react-helmet`,
-    {
-      resolve: `gatsby-plugin-manifest`,
-      options: {
-        name: 'Gatsby + Node.js (TypeScript) API',
-        short_name: 'Gatbsy + Node.js (TypeScript)',
-        start_url: '/',
-        icon: 'src/images/gatsby-icon.png',
-      },
-    },
-  ],
-};
