--- examples/gatsby-functions/package.json
@@ -1,21 +1,0 @@
-{
-  "private": true,
-  "name": "gatsby-starter-default",
-  "version": "1.0.0",
-  "dependencies": {
-    "gatsby": "^2.18.14",
-    "gatsby-image": "^2.0.15",
-    "gatsby-plugin-manifest": "^2.0.5",
-    "gatsby-plugin-react-helmet": "^3.0.0",
-    "react": "^16.5.1",
-    "react-dom": "^16.5.1",
-    "react-helmet": "^5.2.0"
-  },
-  "devDependencies": {
-    "@now/node": "^1.3.0"
-  },
-  "scripts": {
-    "dev": "gatsby develop",
-    "build": "gatsby build"
-  }
-}
