+++ examples/gatsby/.gitignore
