+++ examples/gatsby/.nowignore
@@ -1,2 +1,2 @@
-README.md
 yarn.lock
+README.md
