+++ examples/gatsby/README.md
@@ -2,11 +2,11 @@
 
 # Gatsby Example
 
-This directory is a brief example of a [Gatsby](https://www.gatsbyjs.org/) app that can be deployed with ZEIT Now and zero configuration.
+This directory is a brief example of a [Gatsby](https://www.gatsbyjs.org/) app with [Serverless Functions](https://zeit.co/docs/v2/serverless-functions/introduction) that can be deployed with ZEIT Now and zero configuration.
 
 ## Deploy Your Own
 
-Deploy your own Gatsby project with ZEIT Now.
+Deploy your own Gatsby project, along with Serverless Functions, with ZEIT Now.
 
 [![Deploy with ZEIT Now](https://zeit.co/button)](https://zeit.co/new/project?template=https://github.com/zeit/now-examples/tree/master/gatsby)
 
@@ -14,7 +14,7 @@ _Live Example: https://gatsby.now-examples.now.sh_
 
 ### How We Created This Example
 
-To get started with Gatsby for deployment with ZEIT Now, you can use the [Gatsby CLI](https://www.gatsbyjs.org/docs/gatsby-cli/) to initialize the project:
+To get started with Gatsby on Now, you can use the [Gatsby CLI](https://www.gatsbyjs.org/docs/gatsby-cli/) to initialize the project:
 
 ```shell
 $ gatsby new gatsby-site
@@ -22,7 +22,7 @@ $ gatsby new gatsby-site
 
 ### Deploying From Your Terminal
 
-You can deploy your new Gatsby project with a single command from your terminal using [Now CLI](https://zeit.co/download):
+You can deploy your new Gatsby project, along with [Serverless Functions](https://zeit.co/docs/v2/serverless-functions/introduction), with a single command from your terminal using [Now CLI](https://zeit.co/download):
 
 ```shell
 $ now
