+++ examples/gatsby/api/date.ts
