+++ examples/gatsby/src/pages/index.js
@@ -1,21 +1,56 @@
-import React from "react"
-import { Link } from "gatsby"
+import React, { useEffect, useState } from 'react';
+import { Helmet } from 'react-helmet';
+import '../styles/index.css';
 
-import Layout from "../components/layout"
-import Image from "../components/image"
-import SEO from "../components/seo"
+function Index() {
+  const [date, setDate] = useState(null);
+  useEffect(() => {
+    async function getDate() {
+      const res = await fetch('/api/date');
+      const newDate = await res.text();
+      setDate(newDate);
+    }
+    getDate();
+  }, []);
+  return (
+    <main>
+      <Helmet>
+        <title>Gatsby + Node.js (TypeScript) API</title>
+      </Helmet>
+      <h1>Gatsby + Node.js (TypeScript) API</h1>
+      <h2>
+        Deployed with{' '}
+        <a
+          href="https://zeit.co/docs"
+          target="_blank"
+          rel="noreferrer noopener"
+        >
+          ZEIT Now
+        </a>
+        !
+      </h2>
+      <p>
+        <a
+          href="https://github.com/zeit/now-examples/blob/master/gatsby-functions"
+          target="_blank"
+          rel="noreferrer noopener"
+        >
+          This project
+        </a>{' '}
+        is a <a href="https://www.gatsbyjs.org/">Gatsby</a> app with two
+        directories, <code>/src</code> for static content and <code>/api</code>{' '}
+        which contains a serverless{' '}
+        <a href="https://nodejs.org/en/">Node.js (TypeScript)</a> function. See{' '}
+        <a href="/api/date">
+          <code>api/date</code> for the Date API with Node.js (TypeScript)
+        </a>
+        .
+      </p>
+      <br />
+      <h2>The date according to Node.js (TypeScript) is:</h2>
+      <p>{date ? date : 'Loading date...'}</p>
+    </main>
+  );
+}
 
-const IndexPage = () => (
-  <Layout>
-    <SEO title="Home" />
-    <h1>Hi people</h1>
-    <p>Welcome to your new Gatsby site.</p>
-    <p>Now go build something great.</p>
-    <div style={{ maxWidth: `300px`, marginBottom: `1.45rem` }}>
-      <Image />
-    </div>
-    <Link to="/page-2/">Go to page 2</Link>
-  </Layout>
-)
-
-export default IndexPage
+export default Index;
